ESR class objects

Objects of class ESR, ESR.Spectrum are R6 objects and hence are usable in a more typical object-oriented language than R. For the user the main difference to R's internal S3 and S4 object implementation is that methods for R6 objects belong to the object itself. See details.

# ESR$new(originator, data)
# ESR$set_data(data)
# ESR$set_origin(originator)

Arguments

originator
character name of the function call that created this object

data
data.frame the ESR spectrum to be stored in the

Format

R6Class generator objects

Slots

originator

character Name of the function call that created this object

data

data.frame XY data of the ESR spectrum

Methods

new(originator, data)

Initialize a new object of class ESR

set_data(data)

Sets the data of the object

set_origin(originator)

Sets the originator of the object

Examples

## create an empty ESR base class object
x <- ESR$new()