Fit a dose response curve to ESR data and create a DE-DEmax Plot after Schellmann & Radtke (2001)

This function repeatedly fits a single saturating exponential to ESR data and calculates the De while removing the highest datapoint datapoint after each iteration. It then creates a DE-DEmax-Plot after Schellmann & Radtke (2001) where the DE is plotted against the number of datapoints used for fitting.

calc_DePlateau(input.data, min.DosePoints = 5, fit.weights = "equal",
  model = "EXP", mean.natural = FALSE, show.grid = TRUE,
  output.console = TRUE, ...)

Arguments

input.data
data.frame (required): data frame with two columns for x=Dose, y=ESR.intensity. Optional: a third column containing individual ESR intensity errors can be provided.

min.DosePoints
integer (with default): minimum number of datapoints used for fitting the single saturating exponential.

fit.weights
logical (with default): option whether the fitting is done with equal weights ('equal') or weights proportional to intensity ('prop'). If individual ESR intensity errors are provided, these can be used as weights by using 'error'.

model
character (with default): Currently implemented models: single-saturating exponential ("EXP"), linear ("LIN").

mean.natural
logical (with default): If there are repeated measurements of the natural signal should the mean amplitude be used for fitting?

show.grid
logical (with default): show horizontal grid lines in plots (TRUE/FALSE)

output.console
logical (with default): plot console output (TRUE/FALSE)

...
further arguments passed to plot and par.

Value

Returns terminal output and a plot. In addition, a list is returned containing the following elements:

output

data frame containing the De (datapoints n, De, De.Error, max.Dose).

Details

Fitting methods

For fitting of the dose response curve the nls function with the port algorithm is used. A single saturating exponential in the form of $$y = a*(1-exp(-(x+c)/b))$$ is fitted to the data. Parameters b and c are approximated by a linear fit using lm.

Fit weighting

If 'equal' all datapoints are weighted equally. For 'prop' the datapoints are weighted proportionally by their respective ESR intensity: $$fit.weights = 1/intensity/(sum(1/intensity))$$ If individual errors on ESR intensity are available, choosing 'error' enables weighting in the form of: $$fit.weights = 1/error/(sum(1/error))$$

Note

Fitting of the dose response curve using fit_DRC is largely derived from the plot_GrowthCurve function of the 'Luminescence' package by Kreutzer et al. (2012).

Fitting methods
Currently, only fitting of a single saturating exponential is supported. Fitting of two exponentials or an exponential with a linear term may be implemented in a future release.

References

Galbraith, R.F. & Roberts, R.G., 2012. Statistical aspects of equivalent dose and error calculation and display in OSL dating: An overview and some recommendations. Quaternary Geochronology, 11, pp. 1-27.

Kreutzer, S., Schmidt, C., Fuchs, M.C., Dietze, M., Fischer, M., Fuchs, M., 2012. Introducing an R package for luminescence dating analysis. Ancient TL, 30 (1), pp 1-8.

See also

Examples

# NOT RUN {
##load example data
data(ExampleData.De, envir = environment())

#calculate and plot De-Dmax Plateau
calc_DePlateau(input.data = ExampleData.De,
               min.DosePoints = 5,
               fit.weights = 'prop',
               show.grid = TRUE,
               output.console = FALSE)
# }